requirements.txt
keras==2.11.0
matplotlib==3.6.0
numpy==1.23.5
opencv_python==4.6.0.66
pandas==1.5.0
Pillow==9.3.0
scikit_image==0.19.3
skimage==0.0
streamlit==1.15.1
